Update README.md user review csv files
Android is expanding as an operating system. It has captured around 74% of the total market which is a true indicator of the huge amount of population using Android. Our goal is to help Android developers to know what is the motivating factor for people to download an app. It will also help to find out the factors that affect someone’s decision to download an app. I would like to analyze categories, reviews, prices, ratings, and installs for this purpose and find out how they are interrelated.
There have some Problem Statements that are very useful for finding insight from a given Dataset -

Which category has the most number of installations.
Top 5 apps with the maximum installation.
Top 5 poor app in refrence of instaallation.
How Rating is Important for Application.
Top Category on Play Store.
Find a number of apps in reference to Free or Paid.
Business Objective
The objective of my analysis is to provide insights into Android applications and their categories.
To deep dive into data for the factors of influences on an application, to know why and how certain applications succeed and others.
Finds the key factor that is responsible for app engagement.
Study the detailed information of the app and analyze them.
Finds which attributes are most important for the application.
Also, what is required for an application to be considered as successfully topping the charts?